Prevent search for root directory from searching with just the drive letter
On Windows a path with just the drive letter seems to evaluate to the current working directory which might find the root build script and then overwrite the correct full path to the root build script.  This causes asserts and problems later when functions try and find the root directory.