Use filename as identifier for Executable targets
This allows target prototypes that depend on executables that are built within the same source tree to find the corresponding target in the dependency graph directly.  If executables used a platform independent identifier then depending target prototypes would need a special case for executables that were built in the same dependency graph.