Track active jobs explicitly in Reader as the dequeue size is incorrect
The dequeue pops a job, reducing the dequeue size, before that job is
processed.  That gives a count of, at least, one less thread than is
required.  This leads to deadlocks when child processes pause to wait
for their stdout buffer to be drained while the parent, forge process
waits for the child processes to finish.