Propagate outdated from dependencies for phony targets
Outdated propagation from dependencies was removed and outdated was
calculated purely from relative timestamp difference.  But this doesn't
work for the group targets used to pass multiple source files to one
invocation of Visual C++ on Windows.  Those group targets don't bind to
files and so don't have a timestamp but they should be outdated when
there dependencies are outdated.