Always archive to static libraries that don't exist
A bug was introduced when adding support for transitive static library
dependencies.  In supporting transitive dependencies static libraries
were only archived to if one or more object files were outdated.  If
there were no outdated objects then the static library is touched to
avoid building it again (the static library is then outdated by a
transitive dependency on another static library).  An exception is
thrown when attempting to touch a static library that doesn't exist.