Treat paths on separate drives as being outside the source tree
Absolute paths to different drives are returned as absolute paths from
relative() as that's the only valid representation for them.  This
commit extends the check that a file is within the source tree to cover
this case by required is_relative() of the relative path in addition
to not containing any ".." elements when expressed relative to the
root directory.