Avoid closing write end of pipe when it matches user requested FD
When the write end of the pipe in Executor::pipe() matches the user
requested file descriptor the `dup2()` call is a no-op, the subsequent
`close()` call closes the write end of the pipe.  Writes to the user
request file descriptor write to whichever file ends up being opened to
that file descriptor.