Properly ignore temporary files from Linux file change hooks
The previous fix ignored the mode when O_TMPFILE was passed making those
open() calls likely to fail rather than just be ignored.  This fix
properly opens the file but doesn't log the access for Forge if the
O_TMPFILE flag is passed.

Ignores files opened with O_TMPFILE as they're not accessible from the
file system and are removed when the last process accessing them is
closed.  Thus they can never be part of the dependency graph between
builds.