Allow mismatched types in apply()
Modules that convert some of their settings values, e.g. from tables to
strings, would cause an "attempt to index a string value (local
'destination')" error when the rhs. was still a table but the lhs. was
not.  It was assumed that the lhs. was also still a table.

Fix is to copy source values straight through if either the source or
the destination are not tables.