Only set outdated when timestamp > last write for cleanable targets
When C++ source files depend on generated header files the C++ source
files are constantly marked as outdated.  This happens because the
generated headers have a later timestamp than the C++ source's last
write time.  Forge treats the source file as outdated, this propagates
to the object file, and the object is always rebuilt.

The timestamps of generated headers being later than the C++ source file
doesn't imply that the C++ source file itself is outdated.  Rather the
header's later timestamp should be propagated to the object file and
the timestamp and the object's last write time determine whether or not
the object is outdated.

The fix is to only calculate outdated from timestamp and last write
time for cleanable targets.  Files that are not cleanable are implicitly
not built by Forge.  There is no point setting them to be outdated when
their timestamp is later than their last write time because they can
never be built by Forge.